requirements.txt
Django==3.1.6

djangorestframework==3.12.2
Pillow==7.2.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-logstash==0.4.6
python-slugify==4.0.1
django-redis
pytz==2020.1
greenlet>=0.4.14
gevent==1.4.0
gunicorn==19.9.0
django-rest-swagger==2.2.0
django-environ
hiredis
redis==3.5.3
python-memcached==1.59
pymysql
django-cors-headers
drf-spectacular==0.21.2